refactor add pet useCase
