+++ PetFam.Backend/tests/PetFam.Volunteers.IntegrationTests/PetFam.Volunteers.IntegrationTests.csproj
@@ -15,6 +15,7 @@
         <PackageReference Include="FluentAssertions" Version="7.0.0-alpha.4" />
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="8.0.11" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0"/>
+        <PackageReference Include="NSubstitute" Version="5.3.0" />
         <PackageReference Include="Respawn" Version="6.2.1" />
         <PackageReference Include="Testcontainers.PostgreSql" Version="4.0.0" />
         <PackageReference Include="xunit" Version="2.5.3"/>
@@ -27,6 +28,10 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\src\PetFam.Web\PetFam.Web.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Pets\" />
     </ItemGroup>
 
 </Project>
