--- PetFam.Backend/tests/PetFam.Volunteers.IntegrationTests/sdfgsd.cs
@@ -1,6 +1,0 @@
-namespace PetFam.Volunteers.IntegrationTests;
-
-public class sdfgsd
-{
-    
-}
